+++ test/Tests/ClientTokenManagementTests.cs
@@ -1,12 +1,10 @@
 using System.Text.Json;
-using ClientCredentialTests.Services;
-using Duende.AccessTokenManagement;
 using IdentityModel;
 using IdentityModel.Client;
 using Microsoft.Extensions.DependencyInjection;
 using RichardSzalay.MockHttp;
 
-namespace ClientCredentialTests;
+namespace Duende.AccessTokenManagement.Tests;
 
 public class ClientTokenManagementTests
 {
