--- test/Tests/Services/AppHost.cs
@@ -1,83 +1,0 @@
-﻿// Copyright (c) Duende Software. All rights reserved.
-// See LICENSE in the project root for license information.
-
-using Duende.Bff.Tests.TestFramework;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.IdentityModel.Tokens;
-
-namespace Duende.Bff.Tests.TestHosts
-{
-    public class AppHost : GenericHost
-    {
-        public int? LocalApiStatusCodeToReturn { get; set; }
-
-        public AppHost(string baseAddress = "https://app")
-            : base(baseAddress)
-        {
-            OnConfigureServices += ConfigureServices;
-            OnConfigure += Configure;
-        }
-
-        private void ConfigureServices(IServiceCollection services)
-        {
-            services.AddRouting();
-            services.AddAuthorization();
-
-            services.AddAuthentication(options =>
-                {
-                    options.DefaultScheme = "cookie";
-                    options.DefaultChallengeScheme = "oidc";
-                })
-                .AddCookie("cookie", options =>
-                {
-                    options.Cookie.Name = "web";
-
-                    options.Events.OnSigningOut = async e => { await e.HttpContext.RevokeRefreshTokenAsync(); };
-                })
-                .AddOpenIdConnect("oidc", options =>
-                {
-                    options.Authority = "https://demo.duendesoftware.com";
-
-                    options.ClientId = "interactive.confidential.short";
-                    options.ClientSecret = "secret";
-
-                    options.ResponseType = "code";
-                    options.ResponseMode = "query";
-
-                    options.Scope.Clear();
-                    options.Scope.Add("openid");
-                    options.Scope.Add("profile");
-                    options.Scope.Add("email");
-                    options.Scope.Add("offline_access");
-                    options.Scope.Add("api");
-
-                    options.GetClaimsFromUserInfoEndpoint = true;
-                    options.SaveTokens = true;
-                    options.MapInboundClaims = false;
-
-                    options.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        NameClaimType = "name",
-                        RoleClaimType = "role"
-                    };
-                });
-        }
-
-        private void Configure(IApplicationBuilder app)
-        {
-            app.UseAuthentication();
-
-            app.UseRouting();
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-
-            });
-
-
-        }
-    }
-}
